Update the README
The title of the project has been updated to more accurately describe the functionality of the tool. A table of contents has been added to the README.md for easier navigation of the project’s available documentation. The description of the project has been expanded to give users and future contributors a more accurate picture of the use cases and functionality of the code. Additional information about the tool's language, library and module dependencies, and installation instructions have been added to assist future contributors in reading, running, and updating the project's code.